Add
pan with CTRL key down (support for trackpad / mouse without middle click). Due to a bug in Storybook, this doesn't work well in the doc. To test it in storybook you need to first focus on a INPUT (i.e. a number box in Playground/Stress Test) or the KeyUp won't be registered